+++ Tests/PlayMode/Utilities/Extensions/Tweens/RotationTweenToTests.cs
@@ -6,6 +6,7 @@ using NUnit.Framework;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.TestTools;
+using Utilities.Enums;
 using Utilities.Tweening;
 using Object = UnityEngine.Object;
 
@@ -31,6 +32,8 @@ namespace Tests.Utilities.Extensions.Tweens
 
         [UnityTest]
         public IEnumerator TweenRotationTests(
+            [Values(SPACE.LOCAL, SPACE.WORLD)]
+            SPACE transformSpace,
             [Values(0f, 0.5f, 1f)] float time,
             [Values(CURVE.LINEAR, CURVE.EASE_IN, CURVE.EASE_OUT, CURVE.EASE_IN_OUT)]
             CURVE curve,
@@ -44,7 +47,7 @@ namespace Tests.Utilities.Extensions.Tweens
                 if (time == 0f)
                     LogAssert.Expect(LogType.Error, new Regex(".*Attempting to apply.*"));
 
-                m_transform.TweenTo(target, time, curve, () => { hasCompleted = true; });
+                m_transform.TweenTo(transformSpace, target, time, curve, () => { hasCompleted = true; });
             }
             catch (Exception e)
             {
@@ -55,36 +58,26 @@ namespace Tests.Utilities.Extensions.Tweens
 
             yield return new WaitUntil(() => hasCompleted);
 
-            CustomAssert(m_transform.rotation, target);
+            CustomAssert(transformSpace == SPACE.WORLD ? m_transform.rotation : m_transform.localRotation, target);
         }
-
+        
         [UnityTest]
-        public IEnumerator TweenLocalRotationTests(
+        public IEnumerator TweenRotationCoroutineTests(
+            [Values(SPACE.LOCAL, SPACE.WORLD)]
+            SPACE transformSpace,
             [Values(0f, 0.5f, 1f)] float time,
             [Values(CURVE.LINEAR, CURVE.EASE_IN, CURVE.EASE_OUT, CURVE.EASE_IN_OUT)]
             CURVE curve,
             [ValueSource(nameof(TestTargetRotationValues))]
             Quaternion target)
         {
-            yield return null;
-            bool hasCompleted = false;
-            try
-            {
-                if (time == 0f)
-                    LogAssert.Expect(LogType.Error, new Regex(".*Attempting to apply.*"));
+            if (time == 0f)
+                LogAssert.Expect(LogType.Error, new Regex(".*Attempting to apply.*"));
 
-                m_transform.TweenToLocal(target, time, curve, () => { hasCompleted = true; });
-            }
-            catch (Exception e)
-            {
-                hasCompleted = true;
-                Console.WriteLine(e);
-                throw;
-            }
+            yield return m_transform.TweenToCoroutine(transformSpace, target, time, curve);
 
-            yield return new WaitUntil(() => hasCompleted);
-
-            CustomAssert(m_transform.localRotation, target);
+            
+            CustomAssert(transformSpace == SPACE.WORLD ? m_transform.rotation : m_transform.localRotation, target);
         }
 
         #endregion
@@ -98,8 +91,8 @@ namespace Tests.Utilities.Extensions.Tweens
         public static IEnumerable<Quaternion> TestTargetRotationValues()
         {
             yield return Quaternion.identity;
-            yield return Quaternion.Euler(90, 90, 90);
-            yield return Quaternion.Inverse(Quaternion.Euler(90, 90, 90));
+            yield return Quaternion.Euler(90, -90, 90);
+            yield return Quaternion.Inverse(Quaternion.Euler(90, 90, -90));
         }
 
         private void CustomAssert(Quaternion a, Quaternion b)